rename vector1bit.c -> vectorfloat1bit.c
